Implement package size

+++ lib/src/main/java/com/example/lib/AppPackage.java
@@ -1,13 +1,13 @@
 package com.example.lib;
 
 public class AppPackage {
-    public AppPackage(String name, int size, boolean isSystem) {
+    public AppPackage(String name, long size, boolean isSystem) {
         this.name = name;
         this.size = size;
         this.isSystem = isSystem;
     }
 
     public final String name;
-    public final int size;
+    public final long size;
     public final boolean isSystem;
 }
